chore/#155: controller, domain 안 쓰는 파일 삭제, 안 쓰는 import 삭제, 주석 추가
