+++ src/main/java/com/epi/epilog/app/domain/annotations/ValidOccurenceType.java
@@ -9,6 +9,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+/**
+ * occurenceType 유효성 검사용
+ */
 @Constraint(validatedBy = OccurenceTypeValidator.class)
 @Target({ ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
