+++ src/main/java/com/epi/epilog/app/domain/enums/ActivityLevel.java
@@ -1,5 +1,8 @@
-package com.epi.epilog.app.domain.member;
+package com.epi.epilog.app.domain.enums;
 
+/**
+ * Member 활동 강도
+ */
 public enum ActivityLevel {
     LOW("활동 강도 낮음"),
     MEDIUM("일반적인 활동 강도"),
