+++ src/main/java/com/epi/epilog/app/domain/enums/GenderType.java
@@ -1,0 +1,8 @@
+package com.epi.epilog.app.domain.enums;
+
+/**
+ * Member 성별
+ */
+public enum GenderType {
+    여자, 남자
+}
