+++ src/main/java/com/epi/epilog/app/domain/enums/MealStatus.java
@@ -1,0 +1,8 @@
+package com.epi.epilog.app.domain.enums;
+
+/**
+ * MealChecklist 상태 종류
+ */
+public enum MealStatus {
+    식사함, 건너뜀, 상태없음
+}
