+++ src/main/java/com/epi/epilog/app/domain/enums/MealType.java
@@ -1,5 +1,8 @@
-package com.epi.epilog.app.domain.meal;
+package com.epi.epilog.app.domain.enums;
 
+/**
+ * 식사 종류
+ */
 public enum MealType {
     아침식사, 점심식사, 저녁식사, 간식
 }
