+++ src/main/java/com/epi/epilog/app/domain/enums/MedicationStatus.java
@@ -1,0 +1,8 @@
+package com.epi.epilog.app.domain.enums;
+
+/**
+ * MedicationChecklist 상태 종류
+ */
+public enum MedicationStatus {
+    복용, 미복용, 상태없음
+}
