--- src/main/java/com/epi/epilog/app/domain/enums/RoleType.java
@@ -1,5 +1,0 @@
-package com.epi.epilog.app.domain.enums;
-
-public enum RoleType {
-    모바일, 워치
-}
