+++ src/main/java/com/epi/epilog/app/domain/enums/WeekType.java
@@ -1,5 +1,8 @@
 package com.epi.epilog.app.domain.enums;
 
+/**
+ * Meal, Medication 스케줄러 지정 요일
+ */
 public enum WeekType {
     월,화,수,목,금,토,일
 }
