--- src/main/java/com/epi/epilog/app/domain/meal/MealStatus.java
@@ -1,5 +1,0 @@
-package com.epi.epilog.app.domain.meal;
-
-public enum MealStatus {
-    식사함, 건너뜀, 상태없음
-}
