--- src/main/java/com/epi/epilog/app/domain/medication/MedicationStatus.java
@@ -1,5 +1,0 @@
-package com.epi.epilog.app.domain.medication;
-
-public enum MedicationStatus {
-    복용, 미복용, 상태없음
-}
