--- src/main/java/com/epi/epilog/app/domain/member/GenderType.java
@@ -1,5 +1,0 @@
-package com.epi.epilog.app.domain.member;
-
-public enum GenderType {
-    여자, 남자
-}
