+++ src/main/java/com/epi/epilog/app/dto/MedicationResponseDto.java
@@ -1,6 +1,6 @@
 package com.epi.epilog.app.dto;
 
-import com.epi.epilog.app.domain.medication.MedicationStatus;
+import com.epi.epilog.app.domain.enums.MedicationStatus;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
